immutability; Maybe is a MonadPlus; + mapM; +Unit type; mthen arguments reordered
mzero actually requires to type-parameters: monad and
a type it is applied to. sad.

mthen arguments order was a bogus when monadic actions
had side effects (like printing) -- they were fired in
reverse order.

Unit type is very similar to Void (has a single value),
but as the latter plays a role of Nothing (constructor
forMaybe in Haskell), the former have to be introduced
to play the role of () (crucial for MonadPlus' guard).